Move discovery_booster to ipv8 component
