+++ src/tribler-core/tribler_core/components/gigachannel/community/gigachannel_community.py
@@ -12,7 +12,7 @@ from pony.orm import db_session
 
 from tribler_common.simpledefs import CHANNELS_VIEW_UUID, NTFY
 
-from tribler_core.components.gigachannel.community.discovery_booster import DiscoveryBooster
+from tribler_core.components.ipv8.discovery_booster import DiscoveryBooster
 from tribler_core.components.metadata_store.db.serialization import CHANNEL_TORRENT
 from tribler_core.components.metadata_store.remote_query_community.payload_checker import ObjState
 from tribler_core.components.metadata_store.remote_query_community.remote_query_community import RemoteQueryCommunity
