+++ src/tribler-core/tribler_core/components/ipv8/discovery_booster.py
