+++ src/tribler-core/tribler_core/components/ipv8/tests/test_discovery_booster.py
@@ -1,6 +1,6 @@
 import pytest
 
-from tribler_core.components.gigachannel.community.discovery_booster import DiscoveryBooster
+from tribler_core.components.ipv8.discovery_booster import DiscoveryBooster
 
 TEST_BOOSTER_TIMEOUT_IN_SEC = 10
 TEST_BOOSTER_TAKE_STEP_INTERVAL_IN_SEC = 1
